IPTV stage: move the preview-stage widgets into lib/widgets/iptv/stage/
Moves the self-contained stage widgets out of the 7,569-line
iptv_results_view.dart into three files under lib/widgets/iptv/stage/:

- iptv_stage_floor.dart: IptvStageFloor/State + IptvTuningWavesPainter
  (origin lines 6877-7086)
- iptv_stage_chip.dart:  IptvStageChip/State + IptvTuningBarsPainter
  (origin lines 7088-7238)
- iptv_rail_info.dart:   iptvRailResolutionExp + IptvRailInfo
  (origin lines 74-77 and 7383-7504)

Bodies are unchanged. The only edits are the privacy renames the move
forces (_IptvStageFloor -> IptvStageFloor, _TuningWavesPainter ->
IptvTuningWavesPainter, _TuningBarsPainter -> IptvTuningBarsPainter,
_IptvRailInfo -> IptvRailInfo, _railResExp -> iptvRailResolutionExp), a
super.key on each public constructor for use_key_in_widget_constructors,
and dart format rewrapping the regex declaration. Each widget already took
everything through its constructor, so the new files reference no host
private member. dart:math and dart:ui left the host with their only users.

The shared resolution regex moves rather than being duplicated: the host's
_cockpitIdentity and _IptvFocusStageInfo now read it from the new file.

Preserved quirks: the chip's black-glass 0xB00B0918 fill and onGlass ink
(it floats over live video, so it must not follow the page theme); only
the TUNING state animates, so nothing repaints under a playing video; the
floor's Transform/CustomPaint stay layer-free for the underlay punch-hole.

Pinned by test/iptv_stage_widgets_origin_pin_test.dart (added in the
previous commit, unchanged here), which drives the real IptvResultsView.
